ScopeWeaver-Lab/testing_suite_v0/requirements.txt
google-generativeai
python-dotenv
streamlit
pandas
plotly